Move serializer methods into accounts.UserSerializer
