+++ cadasta/organization/serializers.py
@@ -1,5 +1,5 @@
 from django.utils.text import slugify
-from rest_framework.serializers import ModelSerializer, ValidationError
+from rest_framework.serializers import ModelSerializer
 
 from core.serializers import DetailSerializer
 from accounts.serializers import UserSerializer
@@ -31,26 +31,10 @@ class UserOrganizationSerializer(ModelSerializer):
         read_only_fields = ('id', 'name')
 
 
-class UserAdminSerializer(ModelSerializer):
+class UserAdminSerializer(UserSerializer):
     organizations = UserOrganizationSerializer(many=True, read_only=True)
 
     class Meta:
         model = User
         fields = ('username', 'first_name', 'last_name', 'email',
                   'organizations', 'last_login', 'is_active')
-
-    def validate_username(self, username):
-        instance = self.instance
-        if instance is not None:
-            if (username is not None and
-               username != instance.username):
-                raise ValidationError('Cannot update username')
-        return username
-
-    def validate_last_login(self, last_login):
-        instance = self.instance
-        if instance is not None:
-            if (last_login is not None and
-               last_login != instance.last_login):
-                raise ValidationError('Cannot update last_login')
-        return last_login
